# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "EnableRequest" : {
        "properties" : {
          "allowedProviders" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "applicationName" : {
            "type" : "string"
          },
          "goal" : {
            "type" : "string"
          },
          "maxCostPerRequest" : {
            "format" : "double",
            "type" : "number"
          },
          "maxLatencyMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mode" : {
            "type" : "string"
          },
          "preferredModelClasses" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "FeedbackRequest" : {
        "properties" : {
          "comment" : {
            "type" : "string"
          },
          "requestRef" : {
            "type" : "string"
          },
          "score" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "RoutingDecisionEntity" : {
        "properties" : {
          "chain" : {
            "type" : "string"
          },
          "chosenProvider" : {
            "type" : "string"
          },
          "complexity" : {
            "format" : "double",
            "type" : "number"
          },
          "createdAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "developerId" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mode" : {
            "type" : "string"
          },
          "scoresJson" : {
            "type" : "string"
          },
          "workflowId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Durable execution engine for AI agents: workflows, gateway and console APIs.",
    "title" : "Continuum",
    "version" : "1"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/portal/developer/autopilot/auto-apply" : {
      "put" : {
        "operationId" : "autoApply",
        "parameters" : [ {
          "in" : "query",
          "name" : "value",
          "required" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "autopilot-controller" ]
      }
    },
...
    "/api/portal/developer/autopilot/canary/start/{bundleId}" : {
      "post" : {
        "operationId" : "startCanary",
        "parameters" : [ {
          "in" : "path",
          "name" : "bundleId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "autopilot-controller" ]
      }
    },
    "/api/portal/developer/autopilot/decisions" : {
      "get" : {
        "operationId" : "decisions_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "type" : "object"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "autopilot-controller" ]
      }
    },
...
    "/api/portal/developer/autopilot/enable" : {
      "post" : {
        "operationId" : "enable_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EnableRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "autopilot-controller" ]
      }
    },
    "/api/portal/developer/autopilot/feedback" : {
      "post" : {
        "operationId" : "feedback",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "autopilot-controller" ]
      }
    },
    "/api/portal/developer/autopilot/profile" : {
      "put" : {
        "operationId" : "profile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EnableRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "autopilot-controller" ]
      }
    },
...
    "/api/portal/developer/autopilot/recommendations/{id}/accept" : {
      "post" : {
        "operationId" : "accept",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "autopilot-controller" ]
      }
    },
    "/api/portal/developer/autopilot/recommendations/{id}/reject" : {
      "post" : {
        "operationId" : "reject",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "autopilot-controller" ]
      }
    },
...
    "/api/portal/developer/cascade/decisions" : {
      "get" : {
        "operationId" : "decisions_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 40,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "additionalProperties" : {
                      "type" : "object"
                    },
                    "type" : "object"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "cascade-controller" ]
      }
    },
...
    "/api/routing/decisions" : {
      "get" : {
        "operationId" : "decisions",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/RoutingDecisionEntity"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "routing-controller" ]
      }
    },
...
  }
}